+++ src/main/java/com/chicchoc/sivillage/domain/cart/application/CartService.java
@@ -1,16 +1,8 @@
 package com.chicchoc.sivillage.domain.cart.application;
 
-import com.chicchoc.sivillage.domain.cart.dto.in.CartRequestDto;
-import com.chicchoc.sivillage.domain.cart.dto.out.CartResponseDto;
-import java.util.List;
-import java.util.Optional;
-
 public interface CartService {
 
     // 장바구니 생성
-    void createCart(CartRequestDto cartRequestDto);
-
-    // 장바구니 조회
-    List<CartResponseDto> getCartUuidList(String userUuid);
+    void createCart();
 
 }
