+++ src/main/java/com/chicchoc/sivillage/domain/cart/infrastructure/CartRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CartRepository extends JpaRepository<Cart, Long> {
 
-    List<Cart> findByUserUuid(String userUuid);
+    Cart findByUserUuid(String userUuid);
 }
