+++ src/main/java/com/chicchoc/sivillage/domain/cart/presentation/CartController.java
@@ -1,16 +1,8 @@
 package com.chicchoc.sivillage.domain.cart.presentation;
 
-import com.chicchoc.sivillage.domain.cart.application.CartProductService;
 import com.chicchoc.sivillage.domain.cart.application.CartService;
-import com.chicchoc.sivillage.domain.cart.dto.in.CartRequestDto;
-import com.chicchoc.sivillage.domain.cart.dto.out.CartProductResponseDto;
-import com.chicchoc.sivillage.domain.cart.dto.out.CartResponseDto;
-import com.chicchoc.sivillage.domain.cart.vo.in.CartRequestVo;
-import com.chicchoc.sivillage.domain.cart.vo.out.CartProductResponseVo;
-import com.chicchoc.sivillage.domain.cart.vo.out.CartResponseVo;
 import com.chicchoc.sivillage.global.common.entity.CommonResponseEntity;
 import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -23,12 +15,8 @@ public class CartController {
     private final CartService cartService;
 
     @PostMapping
-    public CommonResponseEntity<Void> createCart(@RequestBody CartRequestVo cartRequestVo) {
-        CartRequestDto cartRequestDto = CartRequestDto.builder()
-                .cartName(cartRequestVo.getCartName())
-                .build();
-
-        cartService.createCart(cartRequestDto);
+    public CommonResponseEntity<Void> createCart() {
+        cartService.createCart();
 
         return new CommonResponseEntity<>(
                 HttpStatus.CREATED,
@@ -37,19 +25,4 @@ public class CartController {
         );
     }
 
-    @GetMapping("/list")
-    public CommonResponseEntity<List<CartResponseVo>> getCartList() {
-        // jwt로 userUuid 가져오기
-        String userUuid = "test1";
-        List<CartResponseDto> cartResponseDto = cartService.getCartUuidList(userUuid);
-        List<CartResponseVo> cartResponseVoList = cartResponseDto.stream()
-                .map(CartResponseDto::toCartResponseVo)
-                .toList();
-
-        return new CommonResponseEntity<>(
-                HttpStatus.OK,
-                "장바구니 불러오기 성공",
-                cartResponseVoList
-        );
-    }
 }
