+++ src/main/java/com/chicchoc/sivillage/domain/cart/vo/in/CartProductRequestVo.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class CartProductRequestVo {
     private String cartUuid;
-    private Long productOrderOptionId;
+    private String productOrderOptionId;
     private int amount;
 }
