--- src/main/java/com/chicchoc/sivillage/domain/cart/vo/in/CartRequestVo.java
@@ -1,9 +1,0 @@
-package com.chicchoc.sivillage.domain.cart.vo.in;
-
-import lombok.Getter;
-
-@Getter
-public class CartRequestVo {
-
-    private String cartName;
-}
